--- a/common/protocol_parser/src/lib.rs
+++ b/common/protocol_parser/src/lib.rs
@@ -2,6 +2,7 @@
 #![feature(hash_raw_entry)]
 extern crate core;
 
+use std::borrow::Cow;
 use std::cmp::Ordering;
 use std::collections::BTreeMap;
 
@@ -11,8 +12,9 @@ use snafu::Snafu;
 use utils::BkdrHasher;
 
 pub type Result<T, E = Error> = std::result::Result<T, E>;
+pub use line_protocol::parser::Error as LineProtocolError;
 
-type NextTagRes<'a> = Result<Option<(Vec<(&'a str, &'a str)>, usize)>>;
+type NextTagRes<'a> = Result<Option<(Vec<(Cow<'a, str>, Cow<'a, str>)>, usize)>>;
 
 pub mod line_protocol;
 pub mod lines_convert;
@@ -34,19 +36,23 @@ pub enum Error {
 #[derive(Debug, PartialEq)]
 pub struct Line<'a> {
     pub hash_id: u64,
-    pub table: &'a str,
-    pub tags: Vec<(&'a str, &'a str)>,
-    pub fields: Vec<(&'a str, FieldValue)>,
+    pub table: Cow<'a, str>,
+    pub tags: Vec<(Cow<'a, str>, Cow<'a, str>)>,
+    pub fields: Vec<(Cow<'a, str>, FieldValue)>,
     pub timestamp: i64,
 }
 
 impl<'a> From<DataPoint<'a>> for Line<'a> {
     fn from(value: DataPoint<'a>) -> Self {
-        let tags = value.tags.into_iter().collect();
-        let fields = vec![("value", FieldValue::F64(value.value))];
+        let tags = value
+            .tags
+            .into_iter()
+            .map(|(s1, s2)| (Cow::Borrowed(s1), Cow::Borrowed(s2)))
+            .collect();
+        let fields = vec![(Cow::Borrowed("value"), FieldValue::F64(value.value))];
         let mut line = Line {
             hash_id: 0,
-            table: value.metric,
+            table: Cow::Borrowed(value.metric),
             tags,
             fields,
             timestamp: value.timestamp,
@@ -68,7 +74,7 @@ impl<'a> Line<'_> {
     pub fn init_hash_id(&mut self) {
         if self.hash_id == 0 {
             self.tags
-                .sort_by(|a, b| -> Ordering { a.0.partial_cmp(b.0).unwrap() });
+                .sort_by(|a, b| -> Ordering { a.0.partial_cmp(&b.0).unwrap() });
 
             let mut hasher = BkdrHasher::new();
             hasher.hash_with(self.table.as_bytes());
@@ -82,9 +88,9 @@ impl<'a> Line<'_> {
     }
 
     pub fn new(
-        measurement: &'a str,
-        tags: Vec<(&'a str, &'a str)>,
-        fields: Vec<(&'a str, FieldValue)>,
+        measurement: Cow<'a, str>,
+        tags: Vec<(Cow<'a, str>, Cow<'a, str>)>,
+        fields: Vec<(Cow<'a, str>, FieldValue)>,
         timestamp: i64,
     ) -> Line<'a> {
         let mut res = Line {
@@ -99,8 +105,8 @@ impl<'a> Line<'_> {
     }
 
     pub fn sort_dedup_and_hash(&mut self) {
-        self.tags.sort_by(|a, b| a.0.cmp(b.0));
-        self.fields.sort_by(|a, b| a.0.cmp(b.0));
+        self.tags.sort_by(|a, b| a.0.cmp(&b.0));
+        self.fields.sort_by(|a, b| a.0.cmp(&b.0));
         self.tags.dedup_by(|a, b| a.0 == b.0);
         self.fields.dedup_by(|a, b| a.0 == b.0);
         self.init_hash_id();
@@ -115,48 +121,6 @@ fn check_pos_valid(buf: &str, pos: usize) -> Result<()> {
         pos,
         content: String::from(buf),
     })
-}
-
-fn next_measurement(buf: &str) -> Option<(&str, usize)> {
-    let mut escaped = false;
-    let mut exists_measurement = false;
-    let (mut tok_begin, mut tok_end) = (0, buf.len());
-    let mut i = 0;
-    for c in buf.chars() {
-        // Measurement begin character
-        if c == '\\' {
-            escaped = true;
-            if !exists_measurement {
-                exists_measurement = true;
-                tok_begin = i;
-            }
-            i += c.len_utf8();
-            continue;
-        }
-        if exists_measurement {
-            // Measurement end character
-            if c == ',' {
-                tok_end = i;
-                break;
-            }
-        } else {
-            // Measurement begin character
-            if c.is_alphanumeric() || c == '_' {
-                exists_measurement = true;
-                tok_begin = i;
-            }
-        }
-        if escaped {
-            escaped = false;
-        }
-
-        i += c.len_utf8();
-    }
-    if exists_measurement {
-        Some((&buf[tok_begin..tok_end], tok_end + 1))
-    } else {
-        None
-    }
 }
 
 fn next_metric(buf: &str) -> Option<(&str, usize)> {
@@ -211,7 +175,7 @@ fn next_tag_set(buf: &str) -> NextTagRes {
     let mut tok_offsets = [0_usize; 3];
     let mut tok_end = 0_usize;
 
-    let mut tag_set: Vec<(&str, &str)> = Vec::new();
+    let mut tag_set = Vec::new();
     let mut i = 0;
     for c in buf.chars() {
         // TagSet begin character
@@ -234,8 +198,8 @@ fn next_tag_set(buf: &str) -> NextTagRes {
             }
             if !escaped && c == ',' {
                 tag_set.push((
-                    &buf[tok_offsets[0]..tok_offsets[1]],
-                    &buf[tok_offsets[2]..i],
+                    Cow::Borrowed(&buf[tok_offsets[0]..tok_offsets[1]]),
+                    Cow::Borrowed(&buf[tok_offsets[2]..i]),
                 ));
                 if buf.len() <= i + 1 {
                     return Ok(None);
@@ -272,223 +236,17 @@ fn next_tag_set(buf: &str) -> NextTagRes {
             return Err(Error::Common {
                 content: format!(
                     "tag missing name in next_tag_set, buf : {}",
-                    &buf.get(0..30).unwrap_or(buf)
+                    buf.get(0..30).unwrap_or(buf)
                 ),
             });
         }
         tag_set.push((
-            &buf[tok_offsets[0]..tok_offsets[1]],
-            &buf[tok_offsets[2]..tok_end],
+            Cow::Borrowed(&buf[tok_offsets[0]..tok_offsets[1]]),
+            Cow::Borrowed(&buf[tok_offsets[2]..tok_end]),
         ));
         Ok(Some((tag_set, tok_end + 1)))
     } else {
         Ok(None)
-    }
-}
-
-type FieldSet<'a> = Vec<(&'a str, FieldValue)>;
-
-fn next_field_set(buf: &str) -> Result<Option<(FieldSet, usize)>> {
-    let mut escaped = false;
-    let mut quoted = false;
-    let mut exists_field_set = false;
-    let mut tok_offsets = [0_usize; 3];
-    let mut tok_end = 0_usize;
-
-    let mut field_set: FieldSet = Vec::new();
-    let mut i = 0;
-    for c in buf.chars() {
-        // TagSet begin character
-        if c == '\\' {
-            escaped = true;
-            if !exists_field_set {
-                exists_field_set = true;
-                tok_offsets[0] = i;
-            }
-            i += c.len_utf8();
-            continue;
-        }
-        if exists_field_set {
-            if !escaped && c == '"' {
-                quoted = !quoted;
-                i += c.len_utf8();
-                continue;
-            }
-
-            if !quoted && c == '=' {
-                tok_offsets[1] = i;
-                if buf.len() <= i + 1 {
-                    return Ok(None);
-                }
-                tok_offsets[2] = i + 1;
-            }
-            if !quoted && c == ',' {
-                field_set.push((
-                    &buf[tok_offsets[0]..tok_offsets[1]],
-                    parse_field_value(&buf[tok_offsets[2]..i])?,
-                ));
-                if buf.len() <= i + 1 {
-                    return Ok(None);
-                }
-                tok_offsets[0] = i + 1;
-            }
-
-            // FieldSet end character
-            if !quoted && (c == ' ' || c == '\n') {
-                tok_end = i;
-                break;
-            }
-        } else {
-            // FieldSet begin character
-            if c.is_alphanumeric() {
-                exists_field_set = true;
-                tok_offsets[0] = i;
-            }
-        }
-        if escaped {
-            escaped = false;
-        }
-
-        i += c.len_utf8();
-    }
-    if exists_field_set {
-        if tok_end == 0 {
-            tok_end = buf.len()
-        }
-        if tok_offsets[0] == tok_offsets[1] {
-            return Err(Error::Common {
-                content: format!(
-                    "field missing name in next_field_set, buf : {}",
-                    &buf.get(0..30).unwrap_or(buf)
-                ),
-            });
-        }
-        field_set.push((
-            &buf[tok_offsets[0]..tok_offsets[1]],
-            parse_field_value(&buf[tok_offsets[2]..tok_end])?,
-        ));
-        Ok(Some((field_set, tok_end + 1)))
-    } else {
-        Ok(None)
-    }
-}
-
-fn parse_field_value(buf: &str) -> Result<FieldValue> {
-    if let Some((_, c)) = buf.chars().enumerate().next() {
-        let ret = match c {
-            ch if ch.is_numeric() => parse_numeric_field(buf, true),
-            '+' => parse_numeric_field(&buf[1..], true),
-            '-' => parse_numeric_field(&buf[1..], false),
-            't' | 'T' => parse_boolean_field(buf, true),
-            'f' | 'F' => parse_boolean_field(buf, false),
-            '"' => parse_string_field(buf),
-            _ => Err(Error::Parse {
-                pos: 0,
-                content: buf.to_string(),
-            }),
-        };
-        return ret;
-    }
-    Ok(FieldValue::F64(1.0))
-}
-
-fn parse_numeric_field(buf: &str, positive: bool) -> Result<FieldValue> {
-    if buf.is_empty() {
-        return Err(Error::Parse {
-            pos: 0,
-            content: buf.to_string(),
-        });
-    }
-    let field_val = match &buf[buf.len() - 1..] {
-        "i" | "I" => {
-            let v = buf[..buf.len() - 1]
-                .parse::<i64>()
-                .map_err(|_e| Error::Parse {
-                    pos: 0,
-                    content: buf.to_string(),
-                })?;
-            FieldValue::I64(if positive { v } else { -v })
-        }
-        "u" | "U" => {
-            if !positive {
-                return Err(Error::Parse {
-                    pos: 0,
-                    content: buf.to_string(),
-                });
-            }
-            let v = buf[..buf.len() - 1]
-                .parse::<u64>()
-                .map_err(|_e| Error::Parse {
-                    pos: 0,
-                    content: buf.to_string(),
-                })?;
-            FieldValue::U64(v)
-        }
-        _ => {
-            let v = buf.parse::<f64>().map_err(|_e| Error::Parse {
-                pos: 0,
-                content: buf.to_string(),
-            })?;
-            FieldValue::F64(if positive { v } else { -v })
-        }
-    };
-
-    Ok(field_val)
-}
-
-fn parse_boolean_field(buf: &str, boolean: bool) -> Result<FieldValue> {
-    const TRUE: &str = "true";
-    const FALSE: &str = "false";
-
-    if buf.len() == 1 {
-        return Ok(FieldValue::Bool(boolean));
-    }
-    let check_iter = if boolean {
-        if buf.len() < TRUE.len() {
-            return Err(Error::Parse {
-                pos: 0,
-                content: buf.to_string(),
-            });
-        }
-        TRUE.chars()
-    } else {
-        if buf.len() < FALSE.len() {
-            return Err(Error::Parse {
-                pos: 0,
-                content: buf.to_string(),
-            });
-        }
-        FALSE.chars()
-    };
-    for (c, check_c) in buf.chars().zip(check_iter) {
-        match c.to_lowercase().next() {
-            Some(ch) => {
-                if ch != check_c {
-                    return Err(Error::Parse {
-                        pos: 0,
-                        content: buf.to_string(),
-                    });
-                }
-            }
-            None => {
-                return Err(Error::Parse {
-                    pos: 0,
-                    content: buf.to_string(),
-                })
-            }
-        }
-    }
-
-    Ok(FieldValue::Bool(boolean))
-}
-
-fn parse_string_field(buf: &str) -> Result<FieldValue> {
-    match &buf[buf.len() - 1..] {
-        "\"" => return Ok(FieldValue::Str(buf[1..buf.len() - 1].as_bytes().to_vec())),
-        _ => Err(Error::Parse {
-            pos: 0,
-            content: buf.to_string(),
-        }),
     }
 }