--- a/common/protocol_parser/src/lines_convert.rs
+++ b/common/protocol_parser/src/lines_convert.rs
@@ -138,7 +138,7 @@ pub enum ColumnData {
 pub fn line_to_batches(lines: &[Line]) -> Result<HashMap<String, MutableBatch>> {
     let mut batches = HashMap::new();
     for line in lines.iter() {
-        let table = line.table;
+        let table = &*line.table;
         let (_, batch) = batches
             .raw_entry_mut()
             .from_key(table)