--- a/query_server/query/src/prom/remote_server.rs
+++ b/query_server/query/src/prom/remote_server.rs
@@ -1,3 +1,4 @@
+use std::borrow::Cow;
 use std::collections::HashMap;
 
 use async_trait::async_trait;
@@ -81,14 +82,25 @@ impl PromRemoteServer for PromRemoteSqlServer {
                     if label.name.eq(METRIC_NAME_LABEL) {
                         table_name = label.value.as_ref()
                     }
-                    (label.name.as_ref(), label.value.as_ref())
+                    (
+                        Cow::Borrowed(label.name.as_ref()),
+                        Cow::Borrowed(label.value.as_ref()),
+                    )
                 })
                 .collect::<Vec<(_, _)>>();
 
             for sample in ts.samples.iter() {
-                let fields = vec![(METRIC_SAMPLE_COLUMN_NAME, FieldValue::F64(sample.value))];
+                let fields = vec![(
+                    Cow::Borrowed(METRIC_SAMPLE_COLUMN_NAME),
+                    FieldValue::F64(sample.value),
+                )];
                 let timestamp = sample.timestamp * 1000000;
-                lines.push(Line::new(table_name, tags.clone(), fields, timestamp));
+                lines.push(Line::new(
+                    Cow::Borrowed(table_name),
+                    tags.clone(),
+                    fields,
+                    timestamp,
+                ));
             }
         }