… Cargo.lock
… common/protocol_parser/src/line_protocol/parser.rs
… query_server/test/cases/sqlancer/select.sql