@zipper-meng 2023-12-20
# Performance

- File: 10G line-protocol.

## Old

| Mode | Cost |
| :--- | :--- |
| single_line | 47695ms |
| single_line | 47620ms |
| multi_line 1000 | 46576ms |
| multi_line 10000 | 49073ms |

## Now

| Mode | Cost |
| :--- | :--- |
| single_line | 101811ms |
| single_line | 102965ms |
| single_line | 110446ms |
| multi_line 1000 | 100097ms |
| multi_line 5000 | 105701ms |
| multi_line 10000 | 105634ms |

## Bench code

### Rust

```rust
use std::fs::File;
use std::io::{BufRead, BufReader};
use std::path::PathBuf;
use std::process::exit;

use protocol_parser::line_protocol::parser::Parser;

fn print_usage() {
    eprintln!(
        r#"Benchmark for line protocol parser.
Usage: benchmark single_line <file path>
    or benchmark multi_line <file path> <line num>
"#
    );
}

fn main() {
    let args = std::env::args().skip(1).collect::<Vec<String>>();
    if args.len() < 2 || args.len() > 3 {
        print_usage();
        exit(1);
    }
    let mode = args[0].as_str();
    eprintln!("- mode: {mode}");
    let path = PathBuf::from(&args[1]);
    eprintln!("- path: {}", path.display());

    let file = File::open(path).unwrap();
    let parser = Parser::new(1);
    let start_time = std::time::Instant::now();
    if mode == "single_line" {
        single_line_mode(file, &parser);
    } else if mode == "multi_line" {
        if args.len() != 3 {
            print_usage();
            exit(1);
        }
        let line_num: u32 = args[2].parse().unwrap();
        eprintln!("- line: {line_num}");
        multi_line_mode(file, &parser, line_num as usize);
    } else {
        print_usage();
        exit(1);
    }
    println!("{}ms", start_time.elapsed().as_millis());
}

fn single_line_mode(file: File, parser: &Parser) {
    let mut reader = BufReader::with_capacity(1024 * 1024 * 128, file);
    let mut line = String::with_capacity(1024);
    loop {
        match reader.read_line(&mut line) {
            Ok(n) => {
                if n == 0 {
                    break;
                }

                match parser.parse(&line) {
                    Ok(_l) => {}
                    Err(e) => panic!("Parse error: {e}"),
                }

                line.clear();
            }
            Err(e) => panic!("Read file error: {e}"),
        }
    }
}

fn multi_line_mode(file: File, parser: &Parser, line_num: usize) {
    let mut reader = BufReader::with_capacity(1024 * 1024 * 16, file);
    let mut buf = String::with_capacity(1024 * 1024);
    let mut ln = 0_usize;
    loop {
        match reader.read_line(&mut buf) {
            Ok(n) => {
                if n == 0 {
                    match parser.parse(&buf) {
                        Ok(_l) => {}
                        Err(e) => panic!("Parse error: {e}"),
                    }
                    break;
                }

                if ln < line_num {
                    ln += 1;
                    continue;
                }

                ln = 0;
                match parser.parse(&buf) {
                    Ok(_l) => {}
                    Err(e) => panic!("Parse error: {e}"),
                }
                buf.clear();
            }
            Err(e) => panic!("Read file error: {e}"),
        }
    }
}
```

### Script

#### Single line mode

```sh
cargo run --release -- single_line seed-123_scale-100_20240101-20240104_interval-1s
```

#### Multi line mode

```sh
cargo run --release -- multi_line seed-123_scale-100_20240101-20240104_interval-1s 1000
```